# docs/03-backend/api-reference/swagger/user-service-api.yaml
openapi: 3.0.3
info:
  title: User Service API
  description: |
    API cho việc quản lý người dùng, authentication, và user profiles trong hệ thống ECommerce.
    
    ## Features
    - User registration và authentication
    - Profile management
    - Password reset
    - Email verification
    - Avatar upload
    
    ## Authentication
    Hầu hết các endpoints yêu cầu JWT token trong Authorization header:
    ```
    Authorization: Bearer <token>
    ```
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@ecommerce.com
servers:
  - url: http://localhost:8081/v1
    description: Local development
  - url: https://api-dev.ecommerce.com/v1
    description: Development environment
  - url: https://api.ecommerce.com/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get list of users
      description: Lấy danh sách users (chỉ admin)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended]
          description: Filter by user status
        - name: roles
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [customer, admin]
          description: Filter by user role(s). Returns users that have any of the supplied roles. Support multiple values (e.g., roles=admin&roles=customer).
        - name: search
          in: query
          schema:
            type: string
          description: Search by username or email
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Lấy thông tin chi tiết user
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Users
      summary: Update user
      description: Cập nhật thông tin user (chỉ owner hoặc admin)
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Xóa user (chỉ admin hoặc owner)
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '204':
          description: User deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token từ /auth/login endpoint
  parameters:
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User ID
      example: 550e8400-e29b-41d4-a716-446655440000
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        example: 'createdAt:desc'
      description: 'Sort format: field:order (e.g., createdAt:desc)'
  schemas:
...
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, inactive, suspended]
        roles:
          type: array
          items:
            type: string
            enum: [customer, admin]
          description: List of roles assigned to the user
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: john.doe@example.com
        username:
          type: string
          example: john_doe
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: https://cdn.ecommerce.com/avatars/john_doe.jpg
        status:
          type: string
          enum: [active, inactive, suspended]
          example: active
        roles:
          type: array
          items:
            type: string
            enum: [customer, admin]
          example: [customer]
        emailVerified:
          type: boolean
          example: true
        birthDate:
          type: string
          format: date
          nullable: true
          example: '1990-01-15'
        bio:
          type: string
          nullable: true
          example: Game enthusiast
        country:
          type: string
          nullable: true
          example: US
        createdAt:
          type: string
          format: date-time
          example: '2025-01-15T10:30:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-10-03T12:00:00.000Z'
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          example: '2025-10-03T11:00:00.000Z'
...
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
        message:
          type: string
          example: User retrieved successfully
        timestamp:
          type: string
          format: date-time
    UsersListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error message
              example: Invalid input data
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time
        trace_id:
          type: string
          description: Request trace ID for debugging
          example: abc-123-def-456
        path:
          type: string
          description: API path that caused error
          example: /v1/users
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Field name that caused error
          example: email
        code:
          type: string
          description: Specific error code for this field
          example: VALIDATION_REQUIRED_FIELD
        message:
          type: string
          description: Error message for this field
          example: Email is required
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        perPage:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 5
        totalItems:
          type: integer
          example: 100
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
  responses:
...
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTH_INVALID_TOKEN
              message: Invalid authentication token
            timestamp: '2025-10-03T12:00:00.000Z'
            trace_id: auth-error-456
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTHZ_FORBIDDEN
              message: You don't have permission to access this resource
            timestamp: '2025-10-03T12:00:00.000Z'
            trace_id: authz-error-789
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RESOURCE_NOT_FOUND
              message: User not found
            timestamp: '2025-10-03T12:00:00.000Z'
            trace_id: not-found-012
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                - field: usernameOrEmail
                  code: VALIDATION_REQUIRED_FIELD
                  message: Username or email is required
                - field: password
                  code: VALIDATION_MIN_LENGTH
                  message: Password must be at least 8 characters
            timestamp: '2025-10-03T12:00:00.000Z'
            trace_id: validation-error-345
...